# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Automify API
  description: |
    API documentation for Automify workflow automation platform.

    ## Authentication
    Most endpoints require authentication via session cookies (managed by Lucia Auth).
    OAuth endpoints use token-based authentication with respective providers.

    ## Rate Limiting
    - **Standard API**: 60 requests/minute
    - **Workflow runs**: 30 runs/minute per user
    - **Auth endpoints**: 5 attempts/minute
    - **Webhooks**: 100 requests/minute per workflow

    Rate limit headers are included in responses:
    - `X-RateLimit-Limit`: Maximum requests allowed
    - `X-RateLimit-Remaining`: Requests remaining in window
    - `X-RateLimit-Reset`: Seconds until window reset
  version: 1.0.0
  contact:
    name: Automify Support
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
...
paths:
...
  /oauth/google/callback:
    get:
      tags:
        - OAuth
      summary: Google OAuth callback
      description: Handles OAuth callback from Google after user authorization
      parameters:
        - in: query
          name: code
          schema:
            type: string
          required: true
          description: Authorization code from Google
        - in: query
          name: state
          schema:
            type: string
          description: Redirect URL encoded in state parameter
      responses:
        "302":
          description: Redirect to connections page or specified redirect URL
        "400":
          description: Missing authorization code
        "500":
          description: OAuth flow failed
...
components:
...
  securitySchemes:
    sessionAuth:
      type: apiKey
      in: cookie
      name: auth_session
      description: Session cookie managed by Lucia Auth